Fix potential Resource is not accessible.

+++ plugins/org.eclipse.acceleo.ide.ui/src/org/eclipse/acceleo/internal/ide/ui/builders/AcceleoCompileOperation.java
@@ -106,19 +106,21 @@ public class AcceleoCompileOperation implements IWorkspaceRunnable {
 			checkCanceled(monitor);
 			monitor.subTask(AcceleoUIMessages.getString(
 					"AcceleoCompileOperation.Task.Clean", files[0].getFullPath().toString())); //$NON-NLS-1$
-			files[i].deleteMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE);
-			files[i].deleteMarkers(IMarker.TASK, true, IResource.DEPTH_INFINITE);
-			files[i].deleteMarkers(AcceleoMarkerUtils.OVERRIDE_MARKER_ID, true, IResource.DEPTH_INFINITE);
-			IPath outputPath = acceleoProject.getOutputFilePath(files[i]);
-			if (outputPath != null) {
-				IFile outputFile = project.getFile(outputPath.removeFirstSegments(1));
-				if (outputFile != null && outputFile.exists()) {
-					try {
-						outputFile.delete(true, monitor);
-					} catch (CoreException e) {
-						// continue
-						// do nothing because it occurs when we have locked the file to write it again
-						AcceleoUIActivator.getDefault().getLog().log(e.getStatus());
+			if (files[i].isAccessible()) {
+				files[i].deleteMarkers(IMarker.PROBLEM, true, IResource.DEPTH_INFINITE);
+				files[i].deleteMarkers(IMarker.TASK, true, IResource.DEPTH_INFINITE);
+				files[i].deleteMarkers(AcceleoMarkerUtils.OVERRIDE_MARKER_ID, true, IResource.DEPTH_INFINITE);
+				IPath outputPath = acceleoProject.getOutputFilePath(files[i]);
+				if (outputPath != null) {
+					IFile outputFile = project.getFile(outputPath.removeFirstSegments(1));
+					if (outputFile != null && outputFile.exists()) {
+						try {
+							outputFile.delete(true, monitor);
+						} catch (CoreException e) {
+							// continue
+							// do nothing because it occurs when we have locked the file to write it again
+							AcceleoUIActivator.getDefault().getLog().log(e.getStatus());
+						}
 					}
 				}
 			}
